1st commit

+++ src/main/java/org/example/service/UserServiceImpl.java
@@ -48,18 +48,30 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void updateUser(User user) {
+    public boolean updateUser(User user) {
+        if (user == null) {
+            return false;
+        }
+
         if (!user.getPassword().equals(getUser(user.getId()).getPassword())) {
             user.setPassword(passwordEncoder.encode(user.getPassword()));
         }
 
         userRepository.save(user);
+
+        return true;
     }
 
     @Override
     @Transactional
-    public void deleteUser(long id) {
+    public boolean deleteUser(long id) {
+        if (getUser(id) == null) {
+            return false;
+        }
+
         userRepository.deleteById(id);
+
+        return true;
     }
 
     @Override
